requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
pydantic==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
sendgrid==6.11.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8